+++ .gitignore
@@ -1,13 +1,7 @@
 */__pycache__/
 __pycache__/
-awsGet.py
-awsManageDynamoDB.ipynb
-config.py
-./keys
-DJuserkey.csv
-rootkey.csv
+aws*
+**/keys
 
 mixed_cookie.txt
-YtVideo.pk
-*.log
-yt_search.json
+**/logs
